Add CellReferenceTests with XML parsing tests
