+++ Tests/CoreXLSXTests/XCTestManifests.swift
@@ -5,6 +5,7 @@ public func allTests() -> [XCTestCaseEntry] {
   return [
     testCase(CoreXLSXTests.allTests),
     testCase(RelationshipsTests.allTests),
+    testCase(CellReferenceTests.allTests),
   ]
 }
 #endif
